# docs/app-api-openapi.yaml
openapi: 3.1.0
info: { title: AmarktAI Thin-App API, version: 1.0.0 }
servers: [{ url: https://network.amarktai.com }]
security: [{ bearerAuth: [] }]
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer } }
...
paths:
...
  /api/v1/artifacts/{id}/file: { get: { summary: Preview or download authorised artifact bytes with Range support, parameters: [{ in: path, name: id, required: true, schema: { type: string } }, { in: query, name: download, schema: { type: string, enum: ['1'] } }, { in: header, name: Range, schema: { type: string } }], responses: { '200': { description: Artifact bytes }, '206': { description: Partial artifact bytes } } } }